# frontend/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "3X-UI Panel API",
    "version": "3.x",
    "description": "Programmatic interface to a 3X-UI panel. Authenticate either by logging in (cookie) or with an API token from Settings → Security → API Token (Bearer). All endpoints under /panel/api/* honour both modes — an API token is a full-admin credential, so treat it like the panel password."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current panel (basePath aware)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API token from Settings → Security → API Token. Send as `Authorization: Bearer <token>`."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "3x-ui",
        "description": "Session cookie set by POST /login. Browser-only."
      }
    },
    "schemas": {
...
      "ProbeResultUI": {
        "properties": {
          "cpuPct": {
            "example": 12.5,
            "type": "number"
          },
          "error": {
            "type": "string"
          },
          "latencyMs": {
            "example": 42,
            "type": "integer"
          },
          "memPct": {
            "example": 45.2,
            "type": "number"
          },
          "panelVersion": {
            "example": "v3.x.x",
            "type": "string"
          },
          "status": {
            "example": "online",
            "type": "string"
          },
          "uptimeSecs": {
            "example": 86400,
            "type": "integer"
          },
          "xrayError": {
            "type": "string"
          },
          "xrayState": {
            "description": "XrayState/XrayError are populated on successful probes even when the node's\nXray core is not healthy. The UI uses them for a distinct \"panel ok, xray failed\" indicator.",
            "type": "string"
          },
          "xrayVersion": {
            "example": "25.10.31",
            "type": "string"
          }
        },
        "required": [
          "cpuPct",
          "error",
          "latencyMs",
          "memPct",
          "panelVersion",
          "status",
          "uptimeSecs",
          "xrayError",
          "xrayState",
          "xrayVersion"
        ],
        "type": "object"
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/panel/api/nodes/mtls/trustCA": {
      "post": {
        "tags": [
          "Nodes"
        ],
        "summary": "Set the CA certificate this panel trusts for incoming node-API client certificates (this panel acting as a node). Paste the managing panel's CA (from nodes/mtls/ca). An empty caCert disables it. A non-empty value must be a PEM certificate. Applied on the next panel restart.",
        "operationId": "post_panel_api_nodes_mtls_trustCA",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "caCert": "-----BEGIN CERTIFICATE-----\n...\n-----END CERTIFICATE-----\n"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/nodes/get/{id}": {
      "get": {
        "tags": [
          "Nodes"
        ],
        "summary": "Fetch a single node by ID.",
        "operationId": "get_panel_api_nodes_get_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Node ID.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/nodes/webCert/{id}": {
      "get": {
        "tags": [
          "Nodes"
        ],
        "summary": "Fetch a node's own web TLS certificate/key file paths (proxied to the node). Used by the inbound form's \"Set Cert from Panel\" so a node-assigned inbound gets paths that exist on the node, not the central panel.",
        "operationId": "get_panel_api_nodes_webCert_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Node ID.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true,
                  "obj": {
                    "webCertFile": "/root/cert/example.com/fullchain.pem",
                    "webKeyFile": "/root/cert/example.com/privkey.pem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/nodes/add": {
      "post": {
        "tags": [
          "Nodes"
        ],
        "summary": "Register a new remote node. Provide its URL, apiToken, and optional remark / allowPrivateAddress flag.",
        "operationId": "post_panel_api_nodes_add",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "name": "de-fra-1",
                "remark": "",
                "scheme": "https",
                "address": "node1.example.com",
                "port": 2053,
                "basePath": "/",
                "apiToken": "abcdef...",
                "enable": true,
                "allowPrivateAddress": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/nodes/provision": {
      "post": {
        "tags": [
          "Nodes"
        ],
        "summary": "Install 3x-ui on a server over SSH, read the generated API token, probe the new panel, and register it as a node. SSH credentials are one-shot request secrets and are not stored.",
        "operationId": "post_panel_api_nodes_provision",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "name": "de-fra-1",
                "sshHost": "203.0.113.10",
                "sshPort": 22,
                "sshUser": "root",
                "sshPassword": "...",
                "sshHostKeySha256": "SHA256:...",
                "sshSkipHostKeyCheck": false,
                "sslMode": "none",
                "panelPort": 54321,
                "webBasePath": "panel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true,
                  "obj": {
                    "node": {
                      "id": 7,
                      "name": "de-fra-1"
                    },
                    "accessUrl": "http://203.0.113.10:54321/panel",
                    "output": [
                      "XUI_API_TOKEN=<redacted>"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/nodes/update/{id}": {
      "post": {
        "tags": [
          "Nodes"
        ],
        "summary": "Replace a node’s connection details. Same body shape as /add.",
        "operationId": "post_panel_api_nodes_update_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Node ID.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "name": "de-fra-1",
                "remark": "",
                "scheme": "https",
                "address": "node1.example.com",
                "port": 2053,
                "basePath": "/",
                "apiToken": "abcdef...",
                "enable": true,
                "allowPrivateAddress": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/nodes/del/{id}": {
      "post": {
        "tags": [
          "Nodes"
        ],
        "summary": "Delete a node. Inbounds bound to it are not auto-migrated.",
        "operationId": "post_panel_api_nodes_del_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Node ID.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/nodes/setEnable/{id}": {
      "post": {
        "tags": [
          "Nodes"
        ],
        "summary": "Pause or resume traffic sync with this node.",
        "operationId": "post_panel_api_nodes_setEnable_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Node ID.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "enable": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/nodes/test": {
      "post": {
        "tags": [
          "Nodes"
        ],
        "summary": "Probe a node without saving it. Uses the body as connection details and returns the same heartbeat snapshot a registered node would have.",
        "operationId": "post_panel_api_nodes_test",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "scheme": "https",
                "address": "node1.example.com",
                "port": 2053,
                "basePath": "/",
                "apiToken": "abcdef..."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {
                      "$ref": "#/components/schemas/ProbeResultUI"
                    }
                  }
                },
                "example": {
                  "success": true,
                  "obj": {
                    "cpuPct": 12.5,
                    "error": "",
                    "latencyMs": 42,
                    "memPct": 45.2,
                    "panelVersion": "v3.x.x",
                    "status": "online",
                    "uptimeSecs": 86400,
                    "xrayError": "",
                    "xrayState": "",
                    "xrayVersion": "25.10.31"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/nodes/certFingerprint": {
      "post": {
        "tags": [
          "Nodes"
        ],
        "summary": "Connect to the node over HTTPS without verifying its certificate and return the leaf certificate's SHA-256 (base64). Used by the Add/Edit Node dialog to fetch and pin a self-signed certificate. Uses the same body as /test.",
        "operationId": "post_panel_api_nodes_certFingerprint",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "scheme": "https",
                "address": "node1.example.com",
                "port": 2053,
                "basePath": "/"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true,
                  "obj": "k3b1...base64-sha256...="
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/nodes/inbounds": {
      "post": {
        "tags": [
          "Nodes"
        ],
        "summary": "Use unsaved node connection details to list the remote inbounds available for selective import.",
        "operationId": "post_panel_api_nodes_inbounds",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "name": "de-fra-1",
                "scheme": "https",
                "address": "node1.example.com",
                "port": 2053,
                "basePath": "/",
                "apiToken": "abcdef..."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true,
                  "obj": [
                    {
                      "tag": "inbound-443",
                      "remark": "VLESS",
                      "protocol": "vless",
                      "port": 443
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/nodes/probe/{id}": {
      "post": {
        "tags": [
          "Nodes"
        ],
        "summary": "Probe an existing node, updating its cached health state.",
        "operationId": "post_panel_api_nodes_probe_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Node ID.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/nodes/updatePanel": {
      "post": {
        "tags": [
          "Nodes"
        ],
        "summary": "Trigger the official panel self-updater on each given node (downloads the latest release and restarts). Only enabled, online nodes are updated; offline/disabled ones are reported as skipped. Set \"dev\": true to move the nodes to the rolling per-commit dev channel instead of the latest stable release. Returns a per-node result list.",
        "operationId": "post_panel_api_nodes_updatePanel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "ids": [
                  1,
                  2,
                  3
                ],
                "dev": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true,
                  "obj": [
                    {
                      "id": 1,
                      "name": "de-1",
                      "ok": true
                    },
                    {
                      "id": 2,
                      "name": "fr-1",
                      "ok": false,
                      "error": "node is offline"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/nodes/history/{id}/{metric}/{bucket}": {
      "get": {
        "tags": [
          "Nodes"
        ],
        "summary": "Aggregated metric history for a node — same shape as /server/history, scoped to one node.",
        "operationId": "get_panel_api_nodes_history_id_metric_bucket",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Node ID.",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "metric",
            "in": "path",
            "required": true,
            "description": "cpu | mem.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bucket",
            "in": "path",
            "required": true,
            "description": "Bucket size in seconds. Allowed: 2, 30, 60, 120, 180, 300.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}